feat: use universal yaml parser

+++ common/global.h
@@ -8,7 +8,7 @@ typedef struct {
   FILE *file;
 } string2file;
 
-extern int external_files_count;
+extern size_t external_files_count;
 extern string2file *external_files;
 
 #endif
